+++ server/src/test/kotlin/io/spine/server/query/QueryingClientSpec.kt
@@ -31,12 +31,12 @@ import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import io.spine.grpc.StreamObservers
 import io.spine.server.BoundedContext
-import io.spine.server.given.counting.NumberStats
-import io.spine.server.given.counting.Range
-import io.spine.server.given.counting.RangeStats
-import io.spine.server.given.counting.command.generateNumbers
-import io.spine.server.given.counting.createCountingContext
-import io.spine.server.given.counting.range
+import io.spine.server.given.context.counting.NumberStats
+import io.spine.server.given.context.counting.Range
+import io.spine.server.given.context.counting.RangeStats
+import io.spine.server.given.context.counting.command.generateNumbers
+import io.spine.server.given.context.counting.createCountingContext
+import io.spine.server.given.context.counting.range
 import io.spine.testing.client.TestActorRequestFactory
 import java.util.Optional.empty
 import org.junit.jupiter.api.AfterAll
