+++ server/src/testFixtures/kotlin/io/spine/server/given/context/counting/CountingContext.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2023, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.given.counting
+package io.spine.server.given.context.counting
 
 import io.spine.core.Subscribe
 import io.spine.server.BoundedContext
@@ -32,9 +32,9 @@ import io.spine.server.BoundedContext.singleTenant
 import io.spine.server.command.Assign
 import io.spine.server.command.SingleCommandAssignee
 import io.spine.server.entity.alter
-import io.spine.server.given.counting.command.GenerateNumbers
-import io.spine.server.given.counting.event.NumberGenerated
-import io.spine.server.given.counting.event.numberGenerated
+import io.spine.server.given.context.counting.command.GenerateNumbers
+import io.spine.server.given.context.counting.event.NumberGenerated
+import io.spine.server.given.context.counting.event.numberGenerated
 import io.spine.server.projection.Projection
 import io.spine.server.projection.ProjectionRepository
 import io.spine.server.route.EventRouting
